Refactor into package and add tooling
