+++ .pre-commit-config.yaml
@@ -1,0 +1,14 @@
+repos:
+  - repo: https://github.com/psf/black
+    rev: 24.4.2
+    hooks:
+      - id: black
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.3.3
+    hooks:
+      - id: ruff
+  - repo: https://github.com/codex-team/clang-format-hooks
+    rev: v12
+    hooks:
+      - id: clang-format
+        files: \.(c|h)$
