+++ Makefile
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 PLUGIN=libspeedtrack.so
 CFLAGS=-Wall -fPIC
 LIBS=$(shell pkg-config --cflags --libs gstreamer-1.0 gstreamer-base-1.0) -lnvds_meta -lsqlite3
